consumer-uploader-service/requirements.txt
requests~=2.32.3
protobuf~=5.29.3
pika~=1.3.2
google-api-python-client~=2.156.0